requirements.txt
requests~=2.32.4
websocket-client~=1.8.0
python-dotenv~=1.1.1
python-binance~=1.0.19